+++ src/components/manager/plannerPage/assignedTab/ManagerAssignedTabContent.tsx
@@ -15,8 +15,8 @@ import Stack from '@mui/material/Stack'
 import Box from '@mui/material/Box'
 import { EntityId } from '@reduxjs/toolkit'
 import { Days, House } from '@/types/schema'
-import { ShiftAssignmentCard } from '@/sprintFiles/ShiftAssignmentCard'
-import { setSelectedUserId } from '@/sprintFiles/userAssignmentSlice'
+import { ShiftAssignmentCard } from '@/features/userAssignment/cards/ShiftAssignmentCard'
+import { setSelectedUserId } from '@/features/userAssignment/userAssignmentSlice'
 
 const filterOptions: Days[] = [
   'All',
