+++ src/features/userAssignment/SelectedUserComponent.tsx
@@ -2,10 +2,10 @@ import { Box, Typography } from '@mui/material'
 import Grid from '@mui/material/Unstable_Grid2'
 import { EntityId } from '@reduxjs/toolkit'
 import { useDispatch, useSelector } from 'react-redux'
-import { selectUserById, useGetUsersQuery } from '../features/user/userApiSlice'
+import { selectUserById, useGetUsersQuery } from '../user/userApiSlice'
 import { RootState } from '@/store/store'
 import { User } from '@/types/schema'
-import CloseButton from './CloseButton'
+import CloseButton from '@/components/shared/buttons/CloseButton'
 import { selectSelectedUserId, setSelectedUserId } from './userAssignmentSlice'
 import { useEffect } from 'react'
