+++ src/store/store.ts
@@ -3,7 +3,7 @@ import counterReducer from './slices/counterSlice'
 import authReducer from '../features/auth/authSlice'
 import usersReducer from '../features/user/usersSlice'
 import userAvailabilityReducer from '../features/userAvailability/userAvailabilitySlice'
-import userAssignmentReducer from '@/sprintFiles/userAssignmentSlice'
+import userAssignmentReducer from '@/features/userAssignment/userAssignmentSlice'
 // import userShiftPreferencesReducer from '../features/userShiftPreferences/userShiftPreferencesSlice'
 import userShiftPreferencesReducer from '../features/userShiftPreferences/userShiftPreferencesSlice'
 import { apiSlice } from './api/apiSlice'
